backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
aiosqlite
alembic
apscheduler
celery[redis]
redis
python-jose[cryptography]
bcrypt
passlib[bcrypt]
pydantic
pydantic-settings
httpx
python-multipart
aiosmtplib
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-prometheus
prometheus-client
jinja2
pyyaml
sentry-sdk[fastapi,celery]
